+++ src/balancer/selection/mod.rs
@@ -1,2 +1,2 @@
 pub mod cache_rules;
-pub mod selection;
+pub mod select;
